discord/requirements.txt
aiohttp==3.8.1
python-dotenv==0.21.0
discord.py==2.0.1
python-dotenv==0.21.0
ipython
asyncpg
